turn left and right

+++ src/Motor.h
@@ -9,7 +9,7 @@ void MotorInit();
 void MotorStop();
 
 void MotorForward(uint8_t sp = 200);
-void MotorForwardLidar(uint8_t minAngle = 240, uint8_t maxAngle = 300);
+void MotorForwardLidar(uint8_t minAngle = 240, uint16_t maxAngle = 300);
 
 void MotorBack(uint8_t sp = 230);
